requirements.txt
cs50==9.2.4
Flask==2.2.2
Pillow==9.3.0
requests==2.28.1
setuptools==63.2.0
Werkzeug==2.2.2
